Added ChatController and implemented some functionality.
Currently, the ChatController class creates a unique chat-key for the student and teacher and sends the messages to that dynamically created endpoint. I added attributesof chat and message classes but havent added them to database. Also lots of khadshaat about identification and verification of users, will discuss in person. Set up the basic flow of the messaging feature atleast.